pyproject.toml
[project]
name = "crypto-portfolio-telegram-bot"
version = "0.1.0"
description = "Track your crypto assets directly through Telegram"
authors = [
    {name = "whiteyod",email = "whiteyod@icloud.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "aiogram>=3.24.0,<4.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "numpy>=2.4.2,<3.0.0",
    "pandas>=3.0.0,<4.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "requests (>=2.32.5,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
